UT improvement
The delete blob test is actually deleting those blobs from putBlob,
with racing with putBlobs (as we run async) and other thread(as all
threads are using same blobID range for deletion).  Each deleteBlob
should either success or fail with unknownBlob.

Previously we access the e.error() with out checkout !!e (success).

When success , e is folly::unit() and accessing e.error() will causing
SIGABRT.

This is the reason that the UT should not pass but was pasing in the
past. It is very unlikely all deleteBlob get executed before putBlob.

Signed-off-by: Xiaoxi Chen <xiaoxchen@ebay.com>