enable restartability (#145)
in the current implementation , when restarting, repl_dev will not be loaded when restarting meta service, it will be postponed to the the start of raft_repl_service. but when restarting, Pg is recovered when restarting metaservice , which will get the repl_dev from raft_repl_service, this causes an error.

temporary fix by caching all pginfo/shardinfo in memory during metaservice recovery.