Add baseline resync read part in homeobject.
Add read_snapshot_data to go over all shards and blobs
of a PG. If obj_id is zero, send all shards. obj_id
is concatenation of blob sequence number and batch number.
For all other values of obj_id, we send batch of blobs for a shard.
Once all blobs are finished in a shard, we move to next shard_id,
and batch_num is reset to 0. Add LSN in shard metadata to ignore
all reads of shards which are created later that the snapshot LSN.