add rollback logic for blob and shard op
if leader receives a put_blob , it will increase the pending req
counter. and when this put_blob is committed, it will decrease the
counter. however, if rollback happens for this pub_blob , the counter
will not be decreased, which will lead to hung in graceful shutdown. the
same as del_blob and created_pg.