add rollback logic for blob and pg op (#274)
* add rollback logic for blob and shard op

if leader receives a put_blob , it will increase the pending req
counter. and when this put_blob is committed, it will decrease the
counter. however, if rollback happens for this pub_blob , the counter
will not be decreased, which will lead to hung in graceful shutdown. the
same as del_blob and created_pg.

* add check for empty header
There is a racing on follower side between commit of a LSN vs fetch_data. i.e when constructing the  fetch data request, the push_data may arrived and the lsn can proceed to commit.  That will result the header be set to empty buffer.

If this case, the LSN should be set to a valid LSN. Leader should ignore this request.