Fix ut
When using spawn to restart the leader, the leader re-runs the ut and attempts to re-create the PG.
Since the PG has been updated by `replace_member`, `create_pg` failed due to differences in PGs.
This fix ensures `create_pg` is skipped after the leader restart to avoid the issue.